+++ src/input.rs
@@ -85,8 +85,8 @@ pub fn process_records(
             match count {
                 0 => header::process(record, out, h)?,
                 _ => match record {
-                    TES3Object::Dialogue(dial) => dial::process(dial, out, h)?,
-                    TES3Object::DialogueInfo(info) => info::process(info, out, h, cfg)?,
+                    TES3Object::Dialogue(dial) => dial::process(dial, out, h, cfg, log)?,
+                    TES3Object::DialogueInfo(info) => info::process(info, out, h, cfg, log)?,
                     _ => {
                         if h.l.active_dial_id.is_some() {
                             h.l.active_dial_id = None;